(ASC-618) fix nova failure on newton
Prior to this PR, the test uses `--network` argument in `openstack server create` command. The test has been executed successfully in Master(Queens), Pike, and Pike-RC. However, when new CI job of newton is in the pipeline, system test in Newton failed because `openstack server create` in Newton does not recognize the `--network` argument.

Based on Openstack doc, `--network` is a wrapper of `--nic <net-id=network-id, ...>`

This PR does:
1. Replacing argument `--network` by `--nic net-id` so that it would pass on all branches incudling Newton
2. Adding a step of verifying the instance is ACTIVE before shutting it down to avoid the fact that the test is trying to shutdown an non-active instance (building or poweroff) that might cause the test failed