ASC-1035 Ensure authorized key on containers
This commit adds a task to ensure that the ssh key for root on localhost
is added to the containers on the target host. Prior to this commit, running
the molecule converge step failed when the MNAIO environment was
deployed from pre-existing images. This is because the deploy hosts key
is propagated to the VMs in the environment but not the containers.